requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
PySocks==1.7.1
requests==2.32.3
urllib3==2.2.3
